requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.13.1
altgraph==0.17.4
amqp==5.2.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==22.12.0
boto3==1.34.5
boto3-stubs==1.34.5
botocore==1.34.5
botocore-stubs==1.34.5
celery==5.2.7
celery-types==0.13.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.0
contourpy==1.2.0
coverage==7.3.4
cryptography==41.0.7
cycler==0.12.1
decorator==5.1.1
docutils==0.18.1
enforce-typing==1.0.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.78.0
fastapi-pagination==0.9.3
flake8==6.1.0
flake8-noqa==1.3.2
flower==1.0.0
fonttools==4.47.0
freezegun==1.4.0
frozenlist==1.4.1
furo==2022.12.7
grblogtools==2.2.0
gurobipy==9.5.0
gurobipy-stubs==1.0.1.post0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanize==4.9.0
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.18.1
ipywidgets==8.1.1
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kombu==5.3.4
macholib==1.16.3
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
moto==4.2.12
multidict==6.0.4
mypy==0.991
mypy-extensions==1.0.0
numpy==1.26.2
openpyxl==3.1.0
openpyxl-stubs==0.1.25
packaging==23.2
pandas==1.4.4
pandas-stubs==1.2.0.62
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pika==0.13.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
PyAMQP==0.1.0.7
pyarrow==7.0.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.13
pyflakes==3.1.0
Pygments==2.17.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.11
PyNaCl==1.5.0
pyodbc==4.0.39
pyparsing==3.1.1
pytest==7.4.3
pytest-cov==3.0.0
pytest-env==0.6.2
python-box==5.4.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.27
python-multipart==0.0.5
pytz==2023.3.post1
pyupgrade==3.15.0
PyYAML==6.0.1
requests==2.31.0
responses==0.24.1
s3transfer==0.9.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==5.3.0
sphinx-basic-ng==1.0.0b2
sphinx-epytext==0.0.4
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-confluencebuilder==2.2.0
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==1.4.29
sqlalchemy2-stubs==0.0.2a37
stack-data==0.6.3
starlette==0.19.1
tenacity==8.2.3
tokenize-rt==5.2.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
types-awscrt==0.20.0
types-click==7.1.8
types-python-dateutil==2.8.19.14
types-pytz==2021.3.8
types-PyYAML==6.0.12.12
types-s3transfer==0.9.0
types-urllib3==1.26.25.14
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.25.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
widgetsnbextension==4.0.9
XlsxWriter==3.1.9
xmltodict==0.13.0
yarl==1.9.4
zipp==3.17.0
